+++ src/ThemeContext.js
@@ -1,8 +1,9 @@
 import {createContext} from 'react'
 
-
 export const ThemeContext = createContext({
     isDarkMode: false,
     toggleTheme: () => {}
   })
+
+
