+++ src/components/TodoListItem.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './static/App.module.css'
+import styles from '../static/App.module.css'
 
 
 export default function TodoListItem({todo, onRemoveTodo}) {
@@ -10,7 +10,7 @@ export default function TodoListItem({todo, onRemoveTodo}) {
             <div className={styles.list}> 
                 <span>{todo.title} </span>
             <button
-            className={styles.buttonRemove}
+             className={styles.buttonRemove}
              type="button"
              onClick={() => 
              onRemoveTodo(todo)
