+++ src/index.css
@@ -1,26 +1,10 @@
-/* body {
-    margin: 0;
-    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-      'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-      sans-serif;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    background-color: rgb(82, 82, 82);
-  } */
-  
-/* @media (30em <= width <= 50em) {
-    body {
-        background-color: black;
-      }
-} */
 
-
-.App {
+.light {
   background-color: #ffffff;
   color: #000000;
 }
 
-.App.dark {
+.dark {
   background-color: #000000;
   color: #ffffff;
 }
